support persistence of discovery results (to handle octane disconnections)
