+++ src/main/java/com/hp/application/automation/tools/octane/executor/UftTestDiscoveryQueue.java
@@ -1,0 +1,34 @@
+/*
+ *     Copyright 2017 Hewlett-Packard Development Company, L.P.
+ *     Licensed under the Apache License, Version 2.0 (the "License");
+ *     you may not use this file except in compliance with the License.
+ *     You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *     Unless required by applicable law or agreed to in writing, software
+ *     distributed under the License is distributed on an "AS IS" BASIS,
+ *     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *     See the License for the specific language governing permissions and
+ *     limitations under the License.
+ *
+ */
+
+package com.hp.application.automation.tools.octane.executor;
+
+import com.hp.application.automation.tools.octane.AbstractResultQueueImpl;
+import jenkins.model.Jenkins;
+
+import java.io.File;
+import java.io.IOException;
+
+/**
+ * Queue for uft test discovery before dispatching
+ */
+public class UftTestDiscoveryQueue extends AbstractResultQueueImpl {
+
+    public UftTestDiscoveryQueue() throws IOException {
+        File queueFile = new File(Jenkins.getInstance().getRootDir(), "octane-uft-tests-queue.dat");
+        init(queueFile);
+    }
+}
